% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllCom.R
\name{getAllCom}
\alias{getAllCom}
\title{Community Detection Based on Feature Correlations}
\usage{
getAllCom(X)
}
\arguments{
\item{X}{A numeric matrix where rows represent samples and columns represent features.}
}
\value{
A list where each element is a vector of feature indices belonging to a detected community.
}
\description{
This function identifies communities (clusters) within features using correlation-based clustering.
Features that are highly correlated are grouped together to form communities.
}
\examples{
set.seed(123)
X <- matrix(rnorm(100), 10, 10)
communities <- getAllCom(X)
str(communities)

}
