% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rho_cv_bic.R
\name{select_rho_cv_bic}
\alias{select_rho_cv_bic}
\title{CV + BIC selector via CVglasso & huge::huge.select}
\usage{
select_rho_cv_bic(X, rhos, seed = NULL, print.detail = TRUE)
}
\arguments{
\item{X}{numeric matrix (samples × features)}

\item{rhos}{numeric vector of penalties (e.g. seq(0.1,0.9,by=0.1))}

\item{seed}{optional integer seed for reproducibility}

\item{print.detail}{logical; if TRUE, prints a summary message}
}
\value{
named numeric vector c(rho.sel, MinCVerr)
}
\description{
CV + BIC selector via CVglasso & huge::huge.select
}
