% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StARS_huge.R
\name{StARS_huge}
\alias{StARS_huge}
\title{StARS via huge::huge + huge::huge.select}
\usage{
StARS_huge(X, b, M, nlambda = 50, beta = 0.05, verbose = TRUE)
}
\arguments{
\item{X}{Numeric data matrix (n samples × p features)}

\item{b}{Subsample size for each stability run}

\item{M}{Number of subsampling iterations (rep.num)}

\item{nlambda}{How many λ’s to fit on the path (default 50)}

\item{beta}{Stability threshold (stars.thresh), default 0.05}

\item{verbose}{Print progress if TRUE}
}
\value{
Named vector c(Sel.rho, D_var)
}
\description{
StARS via huge::huge + huge::huge.select
}
