% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBestPara_JGL.R
\name{[.cordbat_best}
\alias{[.cordbat_best}
\title{Subsetting method to drop names when extracting single penalty value}
\usage{
\method{[}{cordbat_best}(x, i, ...)
}
\description{
Prevent name mismatch when doing best["lambda1"] in tests
}
