% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCorrectCoef.R
\name{update.CorrectCoef}
\alias{update.CorrectCoef}
\title{Batch‐effect `a`/`b` updater (R wrapper)}
\usage{
\method{update}{CorrectCoef}(
  X0.glist,
  X1.glist,
  Theta.list,
  a.i,
  b.i,
  penal.ksi,
  penal.gamma,
  print.detail = TRUE
)
}
\arguments{
\item{print.detail}{logical; echo a message?}
}
\value{
List with components `coef.a` and `coef.b
}
\description{
Batch‐effect `a`/`b` updater (R wrapper)
}
